[Organizations] separation of Organization resource
