+++ examples/resources/organization/resource.tf
@@ -1,16 +1,6 @@
 resource "foundry_organization" "example-organization" {
-  name         = "Example organization nmae"
+  name         = "Example organization name"
   description  = "An example organization in Foundry"
   host_name = "example.palantirfoundry.com"
-  organization_members = ["example-user-id", "example-group-id"]
-  organization_roles = [
-    {
-      "role_id" : "organization:example-role",
-      "principal_id" : "example-user-id",
-    },
-    {
-      "role_id" : "organization:example-role",
-      "principal_id" : "example-group-id",
-    },
-  ]
+  initial_administrators = ["example-user-id", "example-group-id"]
 }
