+++ templates/resources/organization.md.tmpl
@@ -8,13 +8,9 @@ description: |-
 
   The following operations are currently supported:
 
-    - Create an organization
-    - Update an organization
-    - Delete an organization
-    - Import an organization
-    - Add or remove an organization's members
-    - Add or remove an organization's roles
-
+    - Create an Organization
+    - Update an Organization
+    - Import an Organization
 ---
 # Organization
