+++ templates/resources/organization_role_assignments.md.tmpl
@@ -1,0 +1,13 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+# template basis poached from https://github.com/hashicorp/terraform-plugin-docs/blob/main/internal/provider/template.go#L246
+page_title: "Palantir Foundry Organization Role Assignments"
+subcategory: "Organizations"
+description: |-
+  Manages an Organization's Role Assignments.
+---
+# Organization Role Assignments
+
+{{ tffile "examples/resources/organization_role_assignments/resource.tf" }}
+
+{{ .SchemaMarkdown | trimspace }}
